pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.4"]
build-backend = "setuptools.build_meta"

[project]
name = "modoboa"
dynamic = ["version"]
authors = [
  { name="Antoine Nguyen", email="tonio@ngyn.org" },
]
description = "Mail hosting made simple"
readme = "README.rst"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django :: 4.2",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: ISC License (ISCL)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Internet :: WWW/HTTP",
]
keywords = ["email"]
license = {file = "LICENSE"}
dependencies = [
	"django>4.0,<5.0",
	"django-ckeditor==6.7.2",
	"django-phonenumber-field[phonenumbers]==8.0.0",
	"django-reversion==5.1.0",
	"django-xforwardedfor-middleware==2.0",
	"django-otp==1.5.4",
	"django-filter",
	"django-rename-app==0.1.7",
	"django-environ",
	"dj-database-url",
	"djangorestframework>=3.14,<3.16",
	"drf-spectacular",
	"uritemplate",
	"django-oauth-toolkit",
	"django-cors-headers",
	"passlib~=1.7.4",
	"bcrypt", # Requires libffi-dev and python-dev
	"asgiref",
	"dnspython==2.7.0",
	"feedparser==6.0.11",
	"fido2==1.2.0",
	"gevent==24.11.1",
	"Pillow",  # Optional dependency for django-ckeditor
	"progressbar33==2.4",
	"python-dateutil",
	"cryptography",
	"pytz",
	"requests",
	"lxml_html_clean",
	"chardet",
	"ovh",
	"oath",
	"redis>=4.2.0rc1",
	"rrdtool>=0.1.10",
	"qrcode",
	"aiosmtplib",
	# PDF credentials
	"reportlab",
	# DMARC
	"tldextract>=2.0.2",
	"defusedxml>=0.6.0",
	"python-magic>=0.4.24",
	# RQ
	"rq>=1.13.0",
	"django-rq",
	# Sieve
	"sievelib>=1.4.1",
        # User
        "drf-nested-routers",
        "vobject",
        "caldav==1.3.9"
]
[project.optional-dependencies]
dev = [
	"django-bower",
	"django-debug-toolbar",
	"pre-commit",
	"black",
	"ruff",
	"django-extensions",
	"Werkzeug",
	"pyOpenSSL"
]
mysql = ["mysqlclient<2.2.8"]
postgresql = ["psycopg[binary]>=3.1"]
test = [
	"argon2-cffi>=16.1.0",
	"factory-boy<3.4.0",
	"httmock==1.4.0",
	"testfixtures==8.3.0",
	"tox",
]
ldap = ["django-auth-ldap<5.0.0"]

[project.scripts]
modoboa-cli = "modoboa:modoboa_admin"

[project.urls]
Homepage = "http://modoboa.org/"
Documentation = "https://modoboa.readthedocs.io/en/latest/"
Repository = "https://github.com/modoboa/modoboa"
Changelog = "https://github.com/modoboa/modoboa/blob/master/CHANGELOG.md"
Issues = "https://github.com/modoboa/modoboa/issues"

[tool.setuptools.dynamic]
version = {attr = "modoboa.version.get_version"}

[tool.ruff]
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q", "UP"]
